Phase 5: SDL2 hosted mode, 3D maze, asteroid, todo, icons, input fixes
SDL2 hosted mode:
- kernel/hosted/hal.c: full SDL2 HAL — display, keyboard, mouse, audio, disk
- kernel/hosted/main.c: SDL2 entry point, 1024x600 window, 60fps
- Arrow keys / backspace / tab / shift+tab / escape injected via SDL_KEYDOWN
- input.key_down() works via PS/2 key_state table from SDL scancodes
- make hosted / make dist for portable Windows exe

Apps:
- initrd/apps/asteroid.lua: Asteroids clone with DDA collision, particles, lives
- initrd/apps/maze3d.lua: first-person DDA raycaster, procedural maze, minimap, compass
- initrd/apps/todo.lua: nested todo list, #headings, - lists, Tab indent, Shift+Tab dedent, debounced autosave, heading sidebar
- initrd/apps/pixel.lua: :icon <name> saves custom desktop icon to /sys/icons/
- initrd/apps/chirp.lua: notes auto-stop at row boundaries (no sustain bleed)
- initrd/apps/terrain.lua: flat black bg + dark gray grid (no checkerboard)

Desktop:
- initrd/sys/main.lua: pixel-art icons for all apps; custom .mpi icon cache; asteroid/todo/maze3d on desktop

Docs:
- initrd/docs/help.txt, chirp.txt, terrain.txt: in-OS help
- README.md: full rewrite with all controls, Lua API, palette, hosted build
- docs/mpi_format.md: :icon command noted
- CHECKLIST.md: SDL2 hosted, new apps, icon system all checked off

Kernel:
- kernel/vfs/vfs.h+c, kernel.c, klua.c: uintptr_t fix for 64-bit pointer safety
- kernel/cpu/io.h: HOSTED no-op guards for inb/outb

+++ kernel/kernel.c
@@ -262,7 +262,7 @@ void kernel_main(uint32_t magic, mb1_info_t *mb) {
                     serial_puts("[DISK] loading VFS from HDD (");
                     serial_hex(img_size / 1024);
                     serial_puts(" KB)\n");
-                    vfs_init((uint32_t)(uintptr_t)buf, img_size);
+                    vfs_init((uintptr_t)buf, img_size);
                 } else {
                     if (buf) kfree(buf);
                     serial_puts("[DISK] read failed, using initrd\n");
